requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.4
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
flask-form==1.0.0
Flask-Login==0.6.2
Flask-Navigation==0.2.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
icalendar==5.0.7
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
